Done with volunteer's table
Added the ajax link to deny the volunteer;
Added the controller to deny the volunteer;
Adjusted 2 fields in the volunteer model definition;
Made the data values readable (i.e. converted true to yes, false to 0,
empty array to the size of 0);
Removed the “accepted agreement” column;
Added “Why Kiva” Bootstrap modal popup, passed the content as the
argument and retrieved it and pasted in the modal on the click event;
HTML-escaped the “Why Kiva” content;
Added the approving volunteers functionality as the hidden form that is
dynamically constructed on the table build - hacky!
Added the denying volunteers functionality - same way;
Added the memory for the last clicked tab (apps vs volunteers) so that
to be able to go back to the same page when redirecting after
approving/denying volunteers - used localStorage;
Add event delegation for the parts of the page loaded dynamically
(modals for “Why Kiva” and hidden forms to approve/deny volunteers)